Classifies names based on gender

Returns an object that classifies inputted names according to gender.

Usage

get_gender(given, family, api_key)

Arguments

given

A vector of given names (i.e. first names).

family

A vector of family names (i.e. surnames or last names).

api_key

A NameSor API Key. This is typically a long string of mixed-case letters and numbers. Get yours at https://namsor.app/

Value

An object that classifies inputted names according to gender.

Author

Charles Crabtree ccrabtr@umich.edu

Examples